refactor/#98 : heatmap 호직을 훅으로 분리한다
